[hand-control] move ctrl modes into an independent file; move sub_pos_object into an independent file.
