+++ aerial_robot_planning/scripts/mapping_control/hand_ctrl_modes.py
@@ -5,121 +5,14 @@
 # @File : object_position_mapping.py
 # @Software: PyCharm
 
-import os
-import sys
-
 import math
-import time
 import rospy
-from functools import wraps
-from std_msgs.msg import UInt8
-from nav_msgs.msg import Odometry
-from abc import ABC, abstractmethod
-from geometry_msgs.msg import PoseStamped, Twist, Vector3, Quaternion, Transform
+
+from geometry_msgs.msg import Twist, Vector3, Quaternion, Transform
 from trajectory_msgs.msg import MultiDOFJointTrajectory, MultiDOFJointTrajectoryPoint
 
-current_path = os.path.abspath(os.path.dirname(__file__))
-if current_path not in sys.path:
-    sys.path.insert(0, current_path)
-
-from util import check_first_data_received, check_topic_subscription
 from pub_mpc_joint_traj import MPCPubJointTraj
-
-##########################################
-# one-to-one mapping of all instantiation classes required.
-##########################################
-
-import rospy
-
-
-class PositionObjectBase(ABC):
-    def __init__(self, object_name: str, topic_name: str, msg_type):
-        """
-        Base class for subscribing to position information.
-        :param topic_name: The name of the topic to subscribe to.
-        :param msg_type: The message type for position data.
-        """
-        self.object_name = object_name
-        self.position = None
-
-        self.subscriber = self._sub_topic(topic_name, msg_type)
-
-        check_position_initialized(self, "position", self.object_name)
-
-    @check_topic_subscription
-    def _sub_topic(self, topic_name, msg_type):
-        rospy.loginfo(f"Subscribed to {topic_name}")
-        return rospy.Subscriber(topic_name, msg_type, self._position_callback, queue_size=3)
-
-    def _position_callback(self, msg):
-        """
-        Callback function to process incoming position messages.
-        Should be implemented by subclasses.
-        :param msg: The message containing position data.
-        """
-        self.position = msg
-
-    def get_position(self):
-        """
-        Return position information from the message.
-        :return: Extracted position data.
-        """
-        return self.position
-
-
-class Hand(PositionObjectBase):
-    def __init__(self):
-        super().__init__(object_name="Hand", topic_name="/hand/mocap/pose", msg_type=PoseStamped)
-
-
-class Arm(PositionObjectBase):
-    def __init__(self):
-        super().__init__(object_name="Arm", topic_name="/arm/mocap/pose", msg_type=PoseStamped)
-
-
-class Drone(PositionObjectBase):
-    def __init__(self, robot_name):
-        super().__init__(object_name="Drone", topic_name=f"/{robot_name}/uav/cog/odom", msg_type=Odometry)
-
-
-class Glove:
-    def __init__(self):
-        self.object_name = "Glove"
-        self.control_mode = None
-        self.control_mode_sub = self._sub_topic("/hand/control_mode", UInt8, self._glove_data_callback)
-
-        self._check_data_initialized()
-
-    @check_topic_subscription
-    def _sub_topic(self, topic_name, msg_type, callback_func):
-        rospy.loginfo(f"Subscribed to {topic_name}")
-        return rospy.Subscriber(topic_name, msg_type, callback_func, queue_size=3)
-
-    def _check_data_initialized(self):
-        """
-        Waits until the position is initialized. Logs a message repeatedly
-        until a valid position is received.
-        """
-        while not rospy.is_shutdown() and self.control_mode is None:
-            rospy.loginfo(f"Waiting for {self.object_name}'s data...")
-            rospy.sleep(0.2)
-        if self.control_mode is not None:
-            rospy.loginfo(f"{self.object_name}'s data received for the first time")
-
-    def _glove_data_callback(self, msg: UInt8):
-        """
-        Callback function to process incoming messages.
-        Should be implemented by subclasses.
-        :param msg: The message containing glove data.
-        """
-        self.control_mode = msg.data
-
-    def get_control_mode(self):
-        """
-        Return position information from the message.
-        :return: Extracted glove data.
-        """
-        return self.control_mode
+from sub_pos_objects import Hand, Arm, Glove
 
 
 ##########################################
@@ -127,11 +20,11 @@ class Glove:
 ##########################################
 class MappingMode(MPCPubJointTraj):
     def __init__(
-        self,
-        robot_name: str,
-        hand: Hand,
-        arm: Arm,
-        control_mode: Glove,
+            self,
+            robot_name: str,
+            hand: Hand,
+            arm: Arm,
+            control_mode: Glove,
     ):
         super().__init__(robot_name=robot_name, node_name="mapping_mode_traj_pub")
         self.hand = hand
@@ -293,11 +186,11 @@ class MappingMode(MPCPubJointTraj):
 ##########################################
 class SphericalMode(MPCPubJointTraj):
     def __init__(
-        self,
-        robot_name: str,
-        hand: Hand,
-        arm: Arm,
-        control_mode: Glove,
+            self,
+            robot_name: str,
+            hand: Hand,
+            arm: Arm,
+            control_mode: Glove,
     ):
         super().__init__(robot_name=robot_name, node_name="spherical_mode_traj_pub")
         self.hand = hand
@@ -417,9 +310,10 @@ class SphericalMode(MPCPubJointTraj):
         self.expected_a_d_distance = max(1, min(20, self.expected_a_d_distance))
 
         target_position = [
-            a_h_unit_vector[i] * self.expected_a_d_distance + getattr(self.hand.position.pose.position, axis)
-            for i, axis in enumerate(["x", "y"])
-        ] + [self.hand.position.pose.position.z]
+                              a_h_unit_vector[i] * self.expected_a_d_distance + getattr(
+                                  self.hand.position.pose.position, axis)
+                              for i, axis in enumerate(["x", "y"])
+                          ] + [self.hand.position.pose.position.z]
 
         multi_dof_joint_traj = MultiDOFJointTrajectory()
         t_has_started = rospy.Time.now().to_sec() - self.start_time
@@ -580,9 +474,10 @@ class CartesianMode(MPCPubJointTraj):
             self.expected_d_target_distance = 0.2
 
         target_position = [
-            o_h_unit_vector[i] * self.expected_d_target_distance + getattr(self.uav_odom.pose.pose.position, axis)
-            for i, axis in enumerate(["x", "y"])
-        ] + [self.hand.position.pose.position.z]
+                              o_h_unit_vector[i] * self.expected_d_target_distance + getattr(
+                                  self.uav_odom.pose.pose.position, axis)
+                              for i, axis in enumerate(["x", "y"])
+                          ] + [self.hand.position.pose.position.z]
 
         target_hand_distance = math.hypot(
             target_position[0] - current_hand_position[0], target_position[1] - current_hand_position[1]
@@ -636,11 +531,11 @@ class CartesianMode(MPCPubJointTraj):
 ##########################################
 class LockMode(MPCPubJointTraj):
     def __init__(
-        self,
-        robot_name: str,
-        hand: Hand,
-        arm: Arm,
-        control_mode: Glove,
+            self,
+            robot_name: str,
+            hand: Hand,
+            arm: Arm,
+            control_mode: Glove,
     ):
         super().__init__(robot_name=robot_name, node_name="lock_mode_traj_pub")
         self.hand = hand
